feat: handled whiteboard deletion

+++ app/api/whiteboard/[id]/route.ts
@@ -10,7 +10,6 @@ export async function PUT(request: NextRequest, { params }: { params: { id: stri
   const user = await supabase.auth.getUser();
 
   if (!user.data.user.id) return notOk('User not found');
-
   if (Number.isNaN(Number(params.id))) return notOk('Invalid ID');
 
   try {
@@ -20,7 +19,7 @@ export async function PUT(request: NextRequest, { params }: { params: { id: stri
     const { data, error } = await supabase
       .from('whiteboard')
       .update(payload)
-      .match({ whiteboard_id: params.id });
+      .match({ whiteboard_id: params.id, owner_id: user.data.user.id });
 
     if (error) {
       console.log("error", error.message)
@@ -37,6 +36,7 @@ export async function PUT(request: NextRequest, { params }: { params: { id: stri
 export async function DELETE(request: NextRequest, { params, }: { params: { id: string }; }) {
   const supabase = createServerComponentClient({ cookies });
   const whiteboard_id = Number(params.id);
+  const { data: { user } } = await supabase.auth.getUser()
 
   // Validate the id
   if (isNaN(whiteboard_id)) {
@@ -47,7 +47,8 @@ export async function DELETE(request: NextRequest, { params, }: { params: { id: 
     const { error } = await supabase
       .from('whiteboard')
       .delete()
-      .eq('whiteboard_id', whiteboard_id);
+      .eq('whiteboard_id', whiteboard_id)
+      .eq('owner_id', user.id);
 
     if (error) {
       throw error;
